bucket-A1: extract bitcoin_family base from src/impl/ltc/ (SAFE-COSMETIC)
Lifts LTC-duplicated header types into src/impl/bitcoin_family/coin/* as
the reusable base; LTC headers become forwarding shims. Pure refactor,
LTC behavior identical (audit classification: SAFE-COSMETIC + SAFE-
ADDITIVE, validated by include closure).

17 files: 7 new (6 bitcoin_family/coin + 1 core/pow.hpp), 10 modified
(8 ltc/coin + 2 sharechain_storage). +969 / -652.

core/pow.hpp pulled in as a transitive dependency of bitcoin_family/coin/
chain_params.hpp; reduces A3's scope to coin_params.hpp only (was 2 files,
now 1). No other A-series sub-PR ordering changes.

Part 1 of 8 bucket-A sub-PRs implementing the multi-coin landing strategy
confirmed 2026-05-24 (see project_multicoin_landing_strategy.md in agent
memory). Cherry-picked from dash-spv-embedded HEAD 8249dc28.
