+++ src/impl/ltc/coin/block.hpp
@@ -1,6 +1,10 @@
 #pragma once
 
+// LTC block types: inherit generic headers from bitcoin_family,
+// extend BlockType with MWEB (MimbleWimble Extension Blocks) support.
+
 #include "transaction.hpp"
+#include <impl/bitcoin_family/coin/base_block.hpp>
 
 #include <core/uint256.hpp>
 #include <core/pack_types.hpp>
@@ -12,75 +16,12 @@ namespace ltc
 namespace coin
 {
 
-struct SmallBlockHeaderType
-{
-    uint64_t m_version {};
-    uint256 m_previous_block{};
-    uint32_t m_timestamp{};
-    uint32_t m_bits{};
-    uint32_t m_nonce{};
+// Import generic header types from bitcoin_family
+using bitcoin_family::coin::SmallBlockHeaderType;
+using bitcoin_family::coin::BlockHeaderType;
 
-    SERIALIZE_METHODS(SmallBlockHeaderType) { READWRITE(VarInt(obj.m_version), obj.m_previous_block, obj.m_timestamp, obj.m_bits, obj.m_nonce); }
-
-    SmallBlockHeaderType() {}
-
-    void SetNull()
-    {
-        m_version = 0;
-        m_previous_block.SetNull();
-        m_timestamp = 0;
-        m_bits = 0;
-        m_nonce = 0;
-    }
-
-    bool IsNull() const
-    {
-        return (m_bits == 0);
-    }
-};
-
-struct BlockHeaderType : SmallBlockHeaderType
-{
-    uint256 m_merkle_root;
-
-    // Full block header uses fixed 4-byte int32 version (not VarInt like SmallBlockHeaderType)
-    template<typename Stream>
-    void Serialize(Stream& s) const {
-        uint32_t version32 = static_cast<uint32_t>(m_version);
-        ::Serialize(s, version32);
-        ::Serialize(s, m_previous_block);
-        ::Serialize(s, m_merkle_root);
-        ::Serialize(s, m_timestamp);
-        ::Serialize(s, m_bits);
-        ::Serialize(s, m_nonce);
-    }
-    template<typename Stream>
-    void Unserialize(Stream& s) {
-        uint32_t version32;
-        ::Unserialize(s, version32);
-        m_version = version32;
-        ::Unserialize(s, m_previous_block);
-        ::Unserialize(s, m_merkle_root);
-        ::Unserialize(s, m_timestamp);
-        ::Unserialize(s, m_bits);
-        ::Unserialize(s, m_nonce);
-    }
-
-    BlockHeaderType() : SmallBlockHeaderType() { }
-
-    void SetNull()
-    {
-        SmallBlockHeaderType::SetNull();
-        m_merkle_root.SetNull();
-    }
-
-    bool IsNull() const
-    {
-        return (m_bits == 0);
-    }
-
-};
-
+// LTC-specific BlockType: extends BlockHeaderType with MWEB support.
+// Dash/BTC/DOGE would use a simpler BlockType without m_mweb_raw.
 struct BlockType : BlockHeaderType
 {
 	std::vector<MutableTransaction> m_txs;
