+++ src/impl/ltc/coin/softfork_check.hpp
@@ -1,39 +1,6 @@
 #pragma once
-
-#include <nlohmann/json.hpp>
-
-#include <set>
-#include <string>
-
-namespace ltc::coin {
-
-/**
- * Populate `out` with all softfork names found in a single getblockchaininfo
- * field value (either the "softforks" or "bip9_softforks" entry).
- *
- * Handles three formats produced by different litecoind / bitcoind versions:
- *   - Array of objects:  [{"id":"segwit",...}, ...]         (modern litecoind)
- *   - Array of strings:  ["segwit", "taproot", ...]         (compact form)
- *   - Object with keys:  {"segwit":{...}, "taproot":{...}}  (BIP9 style)
- */
-inline void collect_softfork_names(const nlohmann::json& value,
-                                   std::set<std::string>& out)
-{
-    if (value.is_array())
-    {
-        for (const auto& item : value)
-        {
-            if (item.is_object() && item.contains("id") && item["id"].is_string())
-                out.insert(item["id"].get<std::string>());
-            else if (item.is_string())
-                out.insert(item.get<std::string>());
-        }
-    }
-    else if (value.is_object())
-    {
-        for (auto it = value.begin(); it != value.end(); ++it)
-            out.insert(it.key());
-    }
-}
-
-} // namespace ltc::coin
+// Forwarding header: generic softfork check moved to bitcoin_family
+#include <impl/bitcoin_family/coin/softfork_check.hpp>
+namespace ltc { namespace coin {
+    using bitcoin_family::coin::collect_softfork_names;
+}}
